% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/control-function-para.R
\name{cf_para}
\alias{cf_para}
\title{Control function method}
\usage{
cf_para(d, family = "gaussian", hp = 0, h = NULL, h_d = NULL)
}
\arguments{
\item{d}{output of the function format_cf()}

\item{family}{c('gaussian','binary'): 'gaussian' for the normal outcome variable; 'binary' for the binary outcome variable.}

\item{hp}{c(0,1): 0 for the scenario where the horizontal pleiotropy is absent and 1 for the scenario where the horizontal pleiotropy is present.}

\item{h}{the function h(delta1), the default is a linear function.}

\item{h_d}{the derivative of function h(delta1).}
}
\value{
the coefficient, corresponding standard error and P-value of every causal function f_j(X).

the F-value and corresponding P-value for hypothesis testing H_0: f = 0 (the causal effect of the exposure on the outcome is zero).
}
\description{
conduct the control function estimation.
}
