% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spmr.R
\name{cf_semipara}
\alias{cf_semipara}
\title{Semi-parametric control function method}
\usage{
cf_semipara(
  d,
  family = NULL,
  g_est = function(d) {
     return(d)
 },
  action = 0,
  h_est = NULL,
  h_d_est = NULL
)
}
\arguments{
\item{d}{output of the function format_cf_semipara().}

\item{family}{c('gaussian','binary'): 'gaussian' for the normal outcome variable; 'binary' for the binary outcome variable.}

\item{g_est}{predefined g function (the causal effect of covariate on the outcome), and the default is linear function.}

\item{action}{0: the relationship between delta2 and delta1 is linear.}

\item{h_est}{the relationship between delta2 and delta1. if action!=0, h_est should be defined.}

\item{h_d_est}{the derivative of function h_est(). If action!=0, h_d_est should be defined.}
}
\value{
the P-value for hypothesis testing H_0: f = 0 (the causal effect of the exposure on the outcome is zero).
}
\description{
conduct the semi-parametric control function estimation.
}
