% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readData.R
\name{dat2list}
\alias{dat2list}
\title{dat2list
Read list object to projection model data file}
\usage{
dat2list(fn)
}
\value{
update from last year's files
}
\description{
dat2list
Read list object to projection model data file
}
