% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/control-function-para.R
\name{format_cf}
\alias{format_cf}
\title{Format data for the control function method}
\usage{
format_cf(
  data,
  iv_col,
  covariate_col = NULL,
  exposure_col,
  outcome_col,
  f,
  g = NULL
)
}
\arguments{
\item{data}{user input data (containing instrumental variable, exposure, outcome, f function, covariate(optional), g function(optional)).}

\item{iv_col}{colnames corresponding to instrumental variables.}

\item{covariate_col}{colnames corresponding to covariates (optional).}

\item{exposure_col}{colname corresponding to exposure.}

\item{outcome_col}{colname corresponding to outcome.}

\item{f}{predefined causal function of exposure on outcome.}

\item{g}{predefined causal function of covariate on outcome (optional).}
}
\value{
a data frame which is suitable for the function cf_para().
}
\description{
Convert user input data into a data format suitable for the function cf_para().
}
