% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spmr.R
\name{format_cf_semipara}
\alias{format_cf_semipara}
\title{Format data for the semi-parametric control function method}
\usage{
format_cf_semipara(
  data,
  iv_col,
  covariate_col = NULL,
  exposure_col,
  outcome_col
)
}
\arguments{
\item{data}{user input data (containing instrumental variable, exposure, outcome, covariate(optional)).}

\item{iv_col}{colnames corresponding to instrumental variables.}

\item{covariate_col}{colnames corresponding to covariates (optional).}

\item{exposure_col}{colname corresponding to exposure.}

\item{outcome_col}{colname corresponding to outcome.}
}
\value{
a data frame which is suitable for the function cf_semipara().
}
\description{
Convert user input data into a data format suitable for the function cf_semipara().
}
