% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/two-stage-prediction.R
\name{format_sp}
\alias{format_sp}
\title{Format data for the two-stage prediction method}
\usage{
format_sp(data, iv_col, covariate_col = NULL, exposure_col, outcome_col, f)
}
\arguments{
\item{data}{user input data (containing instrumental variable, exposure, outcome, causal function (f), covariate (optional)).}

\item{iv_col}{colnames corresponding to instrumental variable.}

\item{covariate_col}{colnames corresponding to covariates.}

\item{exposure_col}{colname corresponding to exposure.}

\item{outcome_col}{colname corresponding to outcome.}

\item{f}{predefined causal function.}
}
\value{
a data frame which is suitable for the function sp_para().
}
\description{
Convert user input data into a data format suitable for the function sp_para().
}
