% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_tier3a_exec_table.R
\name{make_tier3a_exec_table}
\alias{make_tier3a_exec_table}
\title{Takes spm results and produces the values for the Tier3a executive summary table}
\usage{
make_tier3a_exec_table(run_dir, endyr, the_scalar)
}
\arguments{
\item{run_dir}{the directory where the spm output is located}

\item{endyr}{the current year}

\item{the_scalar}{the scalar used in spm.dat for outputting results}
}
\value{
Values for Tier 3a executive summary table
}
\description{
Takes spm results and produces the values for the Tier3a executive summary table
}
\keyword{internal}
