% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-res.R
\name{plotSPMx}
\alias{plotSPMx}
\title{Plot Results from SPM Analysis from a Tibble}
\usage{
plotSPMx(df, alt = 2, thisyr = 2022, mytitle = NULL)
}
\arguments{
\item{df}{A tibble containing the results from `spm_detail.csv`.}

\item{alt}{The alternative to be plotted.}

\item{thisyr}{The current year for the projection model.}

\item{mytitle}{Title for the plot}
}
\value{
A combined plot of the results from the SPM analysis.
}
\description{
This function creates plots from the results of a Stock Production Model (SPM) analysis contained in a tibble.
It generates three separate plots for different metrics and combines them using the `patchwork` package.
}
