% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spmR.R
\name{read_spm_inputs}
\alias{read_spm_inputs}
\title{Parse SPM input files}
\usage{
read_spm_inputs(dirname)
}
\arguments{
\item{dirname}{Directory containing spm.dat and species files.}
}
\value{
A list with spm input components.
}
\description{
Reads spm.dat and associated species files into a structured list.
}
\keyword{internal}
