% Generated by roxygen2 (4.0.1): do not edit by hand
\name{spm_clusters}
\alias{spm_clusters}
\title{SPM: Return the cluster index for a point list}
\usage{
[A] = spm_clusters(L)
}
\arguments{
\item{L}{locations [x y x]' {in voxels} ([3 x m] matrix)}
}
\value{
cluster index or region number ([1 x m] vector)
}
\description{
spm_clusters characterizes a point list of voxel values defined with
their locations (L) in terms of edge, face and vertex connected
subsets, returning a list of indices in A, such that the ith location
belongs to cluster A(i) (using an 18 connectivity scheme).
}

