% Generated by roxygen2 (4.0.1): do not edit by hand
\name{spm_conman}
\alias{spm_conman}
\title{SPM Contrast Manager}
\usage{
spm_conman(action = "ParseCon", cstr, X, STAT = "F")
}
\arguments{
\item{cstr}{string (or cellstr) to evaluate to get contrast(s)}

\item{X}{design matrix}

\item{STAT}{'T' or 'F' (for contrast checking)}

\item{c}{contrast weights matrix}

\item{I}{logical validity indicator: indicates which rows of cellstr(cstr) generated valid contrasts which were included in c}
}
\description{
Contrast manager for spmR.
}

