% Generated by roxygen2 (4.0.1): do not edit by hand
\name{spm_get_lm}
\alias{spm_get_lm}
\title{SPM: Identification of local maxima in 3(or 2)D volume - a compiled routine}
\usage{
INDEX = spm_get_lm(VOL,LIST)
}
\arguments{
\item{VOL}{3(or 2)D volume of statistics (e.g. t or F)}

\item{LIST}{3xn (or 2xn) list of voxel coordinates of}
}
\value{
INDEX        : Index into LIST such that LIST(:,INDEX)
               returns those coordinates that are truly
               local maxima.
}
\description{
Routine that identifies which voxels in a list of coordinates
that are local maxima, and returns a list of indicies into
the coordinate list for those maxima.
}

