% Generated by roxygen2 (4.0.1): do not edit by hand
\name{spm_getSPM}
\alias{spm_getSPM}
\title{SPM: Compute a specified and thresholded SPM/PPM following estimation}
\usage{
FORMAT [SPM,xSPM] = spm_getSPM;
Query SPM in interactive mode.

FORMAT [SPM,xSPM] = spm_getSPM(xSPM);
Query SPM in batch mode. See below for a description of fields that may
be present in xSPM input. Values for missing fields will be queried
interactively.
}
\value{
xSPM      - structure containing SPM, distribution & filtering details
}
\description{
SPM: Compute a specified and thresholded SPM/PPM following estimation
}

