% Generated by roxygen2 (4.0.1): do not edit by hand
\name{spm_global}
\alias{spm_global}
\title{SPM: returns the global mean for a memory mapped volume image - a compiled routine}
\usage{
GX = spm_global(V)
}
\arguments{
\item{V}{memory mapped volume}
}
\value{
mean global activity
}
\description{
spm_global returns the mean counts integrated over all the
slices from the volume

The mean is estimated after discounting voxels outside the object
using a criteria of greater than > (global mean)/8
}

