% Generated by roxygen2 (4.0.1): do not edit by hand
\name{spm_list}
\alias{spm_list}
\title{spm_list}
\usage{
Summary list of local maxima for entire volume of interest:
TabDat = spm_list('List',SPM,hReg,[Num,Dis,Str])

List of local maxima for a single suprathreshold cluster:
TabDat = spm_list('ListCluster',SPM,hReg,[Num,Dis,Str])
}
\arguments{
\item{action}{}

\item{xSPM}{}

\item{Num}{}

\item{Dis}{}
}
\description{
spm_list characterizes SPMs (thresholded at u and k) in terms of
excursion sets (a collection of face, edge and vertex connected
subsets or clusters).  The corrected significance of the results are
based on set, cluster and voxel-level inferences using distributional
approximations from the Theory of Gaussian Fields.  These
distributions assume that the SPM is a reasonable lattice
approximation of a continuous random field with known component field
smoothness.

The p values are based on the probability of obtaining c, or more,
clusters of k, or more, resels above u, in the volume S analysed =
P(u,k,c).  For specified thresholds u, k, the set-level inference is
based on the observed number of clusters C, = P(u,k,C).  For each
cluster of size K the cluster-level inference is based on P(u,K,1)
and for each voxel (or selected maxima) of height U, in that cluster,
the voxel-level inference is based on P(U,0,1).  All three levels of
inference are supported with a tabular presentation of the p values
and the underlying statistic:

Set-level     - c    = number of suprathreshold clusters
              - P    = prob(c or more clusters in the search volume)

Cluster-level - k    = number of voxels in this cluster
              - Pc   = prob(k or more voxels in the search volume)
              - Pu   = prob(k or more voxels in a cluster)
              - Qc   = lowest FDR bound for which this cluster would be
                       declared positive

Peak-level    - T/F  = Statistic upon which the SPM is based
              - Ze   = The equivalent Z score - prob(Z > Ze) = prob(t > T)
              - Pc   = prob(Ze or higher in the search volume)
              - Qp   = lowest FDR bound for which this peak would be
                       declared positive
              - Pu   = prob(Ze or higher at that voxel)

Voxel-level   - Qu   = Expd(Prop of false positives among voxels >= Ze)

x,y,z (mm)    - Coordinates of the voxel

The table is grouped by regions and sorted on the Ze-variate of the
primary maxima.  Ze-variates (based on the uncorrected p value) are the
Z score equivalent of the statistic. Volumes are expressed in voxels.

Clicking on values in the table returns the value to the Matlab
workspace. In addition, clicking on the co-ordinates jumps the
results section cursor to that location. The table has a context menu
(obtained by right-clicking in the background of the table),
providing options to print the current table as a text table, or to
extract the table data to the Matlab workspace.
}
\seealso{
spm_getSPM
}

