% Generated by roxygen2 (4.0.1): do not edit by hand
\name{spm_P_Bonf}
\alias{spm_P_Bonf}
\title{SPM: Returns the corrected P value using Bonferroni}
\usage{
P = spm_P_Bonf(Z,df,STAT,S,n)
}
\arguments{
\item{Z}{- height {minium over n values}}

\item{df}{- [df{interest} df{error}]}

\item{STAT}{- Statistical field
'Z' - Gaussian field
'T' - T - field
'X' - Chi squared field
'F' - F - field}

\item{n}{- number of conjoint SPMs}

\item{S}{- Voxel count}
}
\value{
P     - corrected   P value  - P(STAT > Z)
}
\description{
spm_P_Bonf returns the p-value of Z corrected by the Bonferroni
inequality.

If n > 1 a conjunction probility over the n values of the statistic
is returned
}

