% Generated by roxygen2 (4.0.1): do not edit by hand
\name{spm_sample_vol}
\alias{spm_sample_vol}
\title{SPM: returns voxel values from a memory mapped image - a compiled routine}
\usage{
X = spm_sample_vol(V,x,y,z,hold);
[X,dX,dY,dZ] = spm_sample_vol(V,x,y,z,hold);
Similar to above, except that the derivatives in the three orthogonal
directions are also returned.
}
\arguments{
\item{V}{a memory mapped image volume}

\item{x}{matrix of x coordinates {pixels}}

\item{y}{matrix of y coordinates {pixels}}

\item{z}{matrix of z coordinates {pixels}}

\item{hold}{sets the interpolation method for the resampling.
 0          Zero-order hold (nearest neighbour).
 1          First-order hold (trilinear interpolation).
 2->127     Higher order Lagrange (polynomial) interpolation using
            different holds (second-order upwards).
-127 - -1   Different orders of sinc interpolation.}
}
\value{
output image
}
\description{
SPM: returns voxel values from a memory mapped image - a compiled routine
}

