% Generated by roxygen2 (4.0.1): do not edit by hand
\name{spm_spm}
\alias{spm_spm}
\title{SPM: [Re]ML Estimation of a General Linear Model}
\usage{
FORMAT [SPM] = spm_spm(SPM)
}
\arguments{
\item{SPM}{}
}
\value{
SPM
}
\description{
Required fields of SPM:

xY.VY - nScan x 1 struct array of image handles (see spm_vol)
        Images must have the same orientation, voxel size and data type
      - Any scaling should have already been applied via the image handle
        scalefactors.

xX    - Structure containing design matrix information
      - Required fields are:
        xX.X      - Design matrix (raw, not temporally smoothed)
        xX.name   - cellstr of parameter names corresponding to columns
                    of design matrix
      - Optional fields are:
        xX.K      - cell of session-specific structures (see spm_filter)
                  - Design & data are pre-multiplied by K
                    (K*Y = K*X*beta + K*e)
                  - Note that K should not smooth across block boundaries
                  - defaults to speye(size(xX.X,1))
        xX.W      - Optional whitening/weighting matrix used to give
                    weighted least squares estimates (WLS). If not specified
                    spm_spm will set this to whiten the data and render
                    the OLS estimates maximum likelihood
                    i.e. W*W' = inv(xVi.V).

xVi   - Structure describing intrinsic temporal non-sphericity
      - Required fields are:
        xVi.Vi    - array of non-sphericity components
                  - defaults to {speye(size(xX.X,1))} - i.i.d.
                  - specifying a cell array of constraints (Qi)
                    These constraints invoke spm_reml to estimate
                    hyperparameters assuming V is constant over voxels.
                    that provide a high precise estimate of xX.V
      - Optional fields are:
        xX.V      - Optional non-sphericity matrix.  Cov(e) = sigma^2*V
                    If not specified spm_spm will compute this using
                    a 1st pass to identify significant voxels over which
                    to estimate V.  A 2nd pass is then used to re-estimate
                    the parameters with WLS and save the ML estimates
                    (unless xX.W is already specified).

xM    - Structure containing masking information, or a simple column vector
        of thresholds corresponding to the images in VY [default: -Inf]
      - If a structure, the required fields are:
        xM.TH - nVar x nScan matrix of analysis thresholds, one per image
        xM.I  - Implicit masking (0=>none, 1 => implicit zero/NaN mask)
        xM.VM - struct array of explicit mask image handles
      - (empty if no explicit masks)
              - Explicit mask images are >0 for valid voxels to assess.
              - Mask images can have any orientation, voxel size or data
                type. They are interpolated using nearest neighbour
                interpolation to the voxel locations of the data Y.
      - Note that voxels with constant data (i.e. the same value across
        scans) are also automatically masked out.

swd   - Directory where the output files will be saved [default: pwd]
        If exists, it becomes the current working directory.

In addition, global SPM "defaults" variable is used (see spm_defaults):

stats.<modality>.UFp - critical F-threshold for selecting voxels over
                       which the non-sphericity is estimated (if
                       required) [default: 0.001]

stats.maxres         - maximum number of residual images for smoothness
                       estimation

stats.maxmem         - maximum amount of data processed at a time (in bytes)

modality             - SPM modality {'PET','FMRI','EEG'}
}

