% Generated by roxygen2 (4.0.1): do not edit by hand
\name{spm_SpUtil}
\alias{spm_SpUtil}
\title{SPM: Space matrix utilities}
\usage{
varargout = spm_SpUtil(action,varargin)
}
\description{
spm_SpUtil is a multi-function function containing various utilities
for Design matrix and contrast construction and manipulation. In
general, it accepts design matrices as plain matrices or as space
structures setup by spm_sp.

Many of the space utilities are computed using an SVD of the design
matrix. The advantage of using space structures is that the svd of
the design matrix is stored in the space structure, thereby saving
unnecessary repeated computation of the SVD. This presents a
considerable efficiency gain for large design matrices.

Note that when space structures are passed as arguments is is
assummed that their basic fields are filled in. See spm_sp for
details of (design) space structures and their manipulation.
}

