% Generated by roxygen2 (4.0.1): do not edit by hand
\name{spm_uc}
\alias{spm_uc}
\title{SPM: corrected critical height threshold at a specified significance level}
\usage{
[u] = spm_uc(a,df,STAT,R,n,S)
}
\arguments{
\item{a}{critical probability - {alpha}}

\item{df}{[df{interest} df{residuals}]}

\item{STAT}{Statistical field
('Z' - Gaussian field,
'T' - T - field,
'X' - Chi squared field,
'F' - F - field)}

\item{R}{RESEL Count {defining search volume}}

\item{n}{number of conjoint SPMs}

\item{S}{Voxel count}
}
\value{
critical height {corrected}
}
\description{
spm_uc corrected critical thresholds, using the minimum of different
valid methods.

See the individual methods for details
}
\seealso{
spm_uc_RF

spm_uc_Bonf
}

