% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readData.R
\name{dat2list}
\alias{dat2list}
\title{Convert Data to List}
\usage{
dat2list(fn)
}
\arguments{
\item{fn}{A character string representing the file name to be read.}
}
\value{
A list with numeric data or character strings based on the content of the file.
}
\description{
This function reads data from a file and converts it into a list.
If the data are numeric, it maintains the numeric list. If the data are strings, it returns a character string.
}
\examples{
# Example usage:
# result <- dat2list("datafile.txt")
}
