% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readData.R
\name{list2dat}
\alias{list2dat}
\title{Write List Object to Projection Model Data File}
\usage{
list2dat(D, fn, hdr = "a new file")
}
\arguments{
\item{D}{List object to be written.}

\item{fn}{Filename where the data will be written.}

\item{hdr}{Header text to be included in the file.}
}
\value{
The function does not return a value; it writes to a file.
}
\description{
This function writes a list object to a file formatted for a projection model.
Each element of the list is written with a header.
}
\examples{
# Example usage:
# list2dat(myList, "datafile.dat", "Header for new file")
}
