% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readData.R
\name{list2dat}
\alias{list2dat}
\title{write list object to projection model data file}
\usage{
list2dat(D, fn, hdr = "a new file")
}
\value{
written data file for spm model
}
\description{
write list object to projection model data file
}
