% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-res.R
\name{plotSPM}
\alias{plotSPM}
\title{Plot SPM Data}
\usage{
plotSPM(df, alt = c(1, 3, 5, 7), mytitle = NULL)
}
\arguments{
\item{df}{A dataframe containing the SPM data with columns `Year`, `Alt`, `variable`, and `value`.}

\item{alt}{A vector of alternatives to include in the plot. Default is `c(1, 4, 5, 7)`.}

\item{mytitle}{An optional title for the plot. Default is `NULL`.}
}
\value{
A ggplot object.
}
\description{
This function filters and processes a dataframe, then creates a plot
of the SPM data with error bands and faceting by type and alternative.
}
\examples{
# Example usage:
df <- data.frame(Year = rep(2000:2020, 4),
                 Alt = rep(1:4, each = 21),
                 variable = rep(c("mean_ub", "mean_lb", "mean_mean"), times = 28),
                 value = runif(84, 0, 1))
plotSPM(df)
}
