% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readData.R
\name{print_Tier3_tables}
\alias{print_Tier3_tables}
\title{print_Tier3_tables}
\usage{
print_Tier3_tables(df, modname = "base", stock = "BSAI Atka mackerel")
}
\arguments{
\item{Dataframe}{(spm_detail.csv)}
}
\value{
formatted table
}
\description{
print_Tier3_tables
}
