% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runSPM.R
\name{runSPM}
\alias{runSPM}
\title{Run SPM Analysis in a Specific Directory}
\usage{
runSPM(dirname, ctrl = NULL)
}
\arguments{
\item{dirname}{A string specifying the directory in which to run the SPM analysis.}

\item{ctrl}{Optional control settings for the SPM analysis. If NULL, default settings are used.}
}
\value{
A data frame containing the results from `spm_detail.csv`.
}
\description{
This function runs a Stock Production Model (SPM) analysis in the specified directory.
The function changes the working directory to `dirname`, runs the SPM analysis, and then reads the results from `spm_detail.csv`.
It returns to the original working directory after completing the analysis.
}
\examples{
runSPM("examples/atka")
}
